+++ libs/cart/src/selectors/cart/cart.selector.spec.ts
@@ -5,11 +5,11 @@ import { cold } from 'jasmine-marbles';
 import { DaffCart } from '@daffodil/cart';
 import { DaffCartFactory } from '@daffodil/cart/testing';
 
-import { DaffCartLoadSuccess, DaffCartPlaceOrderSuccess } from '../actions/public_api';
-import { daffCartReducers, DaffCartReducersState } from '../reducers/public_api';
-import { getDaffCartSelectors } from './cart.selector';
-import { DaffCartErrorType } from '../reducers/cart-error-type.enum';
-import { DaffCartErrors } from '../reducers/cart-errors.type';
+import { DaffCartLoadSuccess, DaffCartPlaceOrderSuccess } from '../../actions/public_api';
+import { daffCartReducers, DaffCartReducersState } from '../../reducers/public_api';
+import { getCartSelectors } from './cart.selector';
+import { DaffCartErrorType } from '../../reducers/cart-error-type.enum';
+import { DaffCartErrors } from '../../reducers/cart-errors.type';
 
 describe('Cart | Selector | Cart', () => {
   let store: Store<DaffCartReducersState>;
@@ -44,14 +44,8 @@ describe('Cart | Selector | Cart', () => {
 		selectCartShippingInformation,
 		selectCartAvailableShippingMethods,
 		selectCartAvailablePaymentMethods,
-    selectIsCartEmpty,
-
-    selectCartOrderState,
-    selectCartOrderLoading,
-    selectCartOrderErrors,
-    selectCartOrderValue,
-    selectCartOrderId
-	} = getDaffCartSelectors();
+    selectIsCartEmpty
+	} = getCartSelectors();
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -295,42 +289,6 @@ describe('Cart | Selector | Cart', () => {
     it('selects whether the cart is empty', () => {
       const selector = store.pipe(select(selectIsCartEmpty));
       const expected = cold('a', {a: cart.items.length === 0});
-
-      expect(selector).toBeObservable(expected);
-    });
-  });
-
-  describe('selectCartOrderLoading', () => {
-    it('selects whether the place order operation is in progress', () => {
-      const selector = store.pipe(select(selectCartOrderLoading));
-      const expected = cold('a', {a: false});
-
-      expect(selector).toBeObservable(expected);
-    });
-  });
-
-  describe('selectCartOrderErrors', () => {
-    it('selects the errors associated with place order', () => {
-      const selector = store.pipe(select(selectCartOrderErrors));
-      const expected = cold('a', {a: []});
-
-      expect(selector).toBeObservable(expected);
-    });
-  });
-
-  describe('selectCartOrderValue', () => {
-    it('selects the order object', () => {
-      const selector = store.pipe(select(selectCartOrderValue));
-      const expected = cold('a', {a: {id: orderId}});
-
-      expect(selector).toBeObservable(expected);
-    });
-  });
-
-  describe('selectCartOrderId', () => {
-    it('selects the ID of the order object', () => {
-      const selector = store.pipe(select(selectCartOrderId));
-      const expected = cold('a', {a: orderId});
 
       expect(selector).toBeObservable(expected);
     });
